Separated dimensions and dimensional arithmetic into seperate modules.
