+++ Numeric/Units/Dimensional/DK/Dimensions.hs
@@ -1,0 +1,11 @@
+-- | Provides both term-level and type-level representations for physical dimensions in 
+-- a single import for convenience.
+module Numeric.Units.Dimensional.DK.Dimensions
+(
+  module Numeric.Units.Dimensional.DK.Dimensions.TermLevel,
+  module Numeric.Units.Dimensional.DK.Dimensions.TypeLevel
+)
+where
+
+import Numeric.Units.Dimensional.DK.Dimensions.TermLevel
+import Numeric.Units.Dimensional.DK.Dimensions.TypeLevel
