+++ dimensional-dk.cabal
@@ -34,7 +34,10 @@ Exposed-Modules:     Numeric.Units.Dimensional.DK,
                      Numeric.Units.Dimensional.DK.Prelude,
                      Numeric.Units.Dimensional.DK.Quantities,
                      Numeric.Units.Dimensional.DK.SIUnits,
-                     Numeric.Units.Dimensional.DK.NonSI
+                     Numeric.Units.Dimensional.DK.NonSI,
+                     Numeric.Units.Dimensional.DK.Dimensions,
+                     Numeric.Units.Dimensional.DK.Dimensions.TermLevel,
+                     Numeric.Units.Dimensional.DK.Dimensions.TypeLevel
 Extra-source-files:  README.md,
                      Test.hs,
                      Numeric/Units/Dimensional/DK/Test.hs,
